First version of a script that parses metrics from Alfano 6
More specifically, in the Alfano app you need to select the
archive/backup option.

This is based on a script where someone wanted to get Assetto Corsa
data into Track Attack for analyzing. It turns out Track Attack knows
how to read AIM (Mychron) CSV files. In the end I didn't get it to
work in Track Attack, and the error message is uselussly opaque...
So the result is that I now have something that can read Alfano data
files, and write AIM CSV files.